pyproject.toml
[project]
name = "rl-emails"
version = "2.0.0"
description = "Email ML pipeline for analyzing Gmail exports and predicting email priority/actions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Database
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",

    # Data processing
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Email parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # LLM APIs
    "openai>=1.0.0",
    "litellm>=1.0.0",
    "tiktoken>=0.5.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",

    # Validation
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.5.0",

    # Type stubs
    "types-tqdm>=4.66.0",
    "types-beautifulsoup4>=4.12.0",
    "types-psycopg2>=2.9.0",
]

surrealdb = [
    "surrealdb>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
rl-emails = "rl_emails.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rl_emails"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

# Ignore missing imports for packages without stubs
[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "sklearn.*",
    "asyncpg.*",
    "surrealdb.*",
    "alembic.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["rl_emails", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/rl_emails"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
partial_branches = [
    "pragma: no branch",
    "for .* in .*:",
    "any\\(",
]
